+++ src/components/pages/AppDetail/AppDetailContentDetails.tsx
@@ -26,15 +26,15 @@ import {
   BackButton,
 } from '@catena-x/portal-shared-components'
 import { useTranslation } from 'react-i18next'
-import AppDetailHeader from './components/AppDetailHeader'
+import AppDetailHeader from './AppDetailHeader'
 import AppDetailPrivacy from './components/AppDetailPrivacy'
-import AppDetailDocuments from './components/AppDetailDocuments'
-import AppDetailProvider from './components/AppDetailProvider'
-import AppDetailTags from './components/AppDetailTags'
+import AppDetailDocuments from './AppDetailDocuments'
+import AppDetailProvider from './AppDetailProvider'
+import AppDetailTags from './AppDetailTags'
 import type { AppDetails } from 'features/apps/types'
 import './style.scss'
 import CommonService from 'services/CommonService'
-import AppDetailTechUserSetup from './components/AppDetailTechUserSetup'
+import AppDetailTechUserSetup from './AppDetailTechUserSetup'
 import { Box } from '@mui/material'
 import { useNavigate } from 'react-router-dom'
 import { PAGES } from 'types/Constants'
