+++ src/components/pages/AppDetail/AppDetailHeader/style.scss
