+++ src/components/pages/AppDetail/AppDetailPrivacy/index.tsx
@@ -22,7 +22,7 @@ import { useTranslation } from 'react-i18next'
 import { Typography } from '@catena-x/portal-shared-components'
 import { uniqueId } from 'lodash'
 import { Apartment, Person, LocationOn, Web, Info } from '@mui/icons-material'
-import './AppDetailPrivacy.scss'
+import './style.scss'
 import type { AppDetails } from 'features/apps/types'
 import { PrivacyPolicyType } from 'features/adminBoard/adminBoardApiSlice'
 import '../../style.scss'
