+++ src/components/pages/AppDetail/AppDetailPrivacy/style.scss
