+++ src/components/pages/AppDetail/AppDetailTags/index.tsx
@@ -20,7 +20,7 @@
 
 import { Chip, Typography } from '@catena-x/portal-shared-components'
 import type { AppDetails } from 'features/apps/types'
-import './AppDetailTags.scss'
+import './style.scss'
 import '../../style.scss'
 import { useTranslation } from 'react-i18next'
