+++ src/components/pages/AppDetail/AppDetailTags/style.scss
