+++ src/components/pages/AppMarketplace/AppListGroupView/style.scss
